crypto: move OlmMachine tests to a separate file
